requirements.txt
absl-py==2.1.0
anyio==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
compress_json==1.0.10
contourpy==1.2.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.3.1
gast==0.4.0
gdown==5.2.0
gmpy2==2.1.2
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
idna==3.7
imageio==2.34.1
imbalanced-learn==0.12.3
imblearn==0.0
importlib_metadata==7.1.0
importlib-resources==6.1.1
intel-openmp==2021.4.0
ipykernel==6.28.0
ipython==8.15.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.6
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab_widgets==3.0.10
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-tuner==1.4.7
kiwisolver==1.4.4
kt-legacy==1.0.5
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.3.0
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mkl==2021.4.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
notebook==7.0.8
notebook_shim==0.2.3
numba==0.59.1
numexpr==2.8.7
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.7.0.72
opt-einsum==3.3.0
optree==0.11.0
overrides==7.4.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==3.10.0
plot_keras_history==1.1.38
plotly==5.22.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
Pygments==2.15.1
pynndescent==0.5.12
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.30.2
regex==2024.5.10
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.3
sanitize_ml_labels==1.0.51
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.12.2
Send2Trash==1.8.2
setuptools==68.2.2
sip==6.7.12
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
support_developer==1.0.5
sympy==1.12
tbb==2021.12.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-addons==0.22.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==2.2.0
tifffile==2024.5.22
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.4
traitlets==5.7.1
typeguard==2.13.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2023.3
umap-learn==0.5.6
unicodedata2==15.1.0
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==3.0.2
wheel==0.43.0
widgetsnbextension==4.0.10
win-inet-pton==1.1.0
wrapt==1.16.0
zipp==3.18.1
